pyproject.toml
[project]
name = "sefaria-discord-bot"
version = "1.0.0"
description = "Discord bot that integrates with Sefaria API to provide access to Jewish texts"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "discord-py>=2.5.2",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
sefaria-bot = "deploy:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]
